+++ src/server/mod.rs
@@ -14,10 +14,11 @@ use crate::lsp_data::{
     ShowMessageParams, Workspace,
 };
 use crate::server::dispatch::Dispatcher;
-pub use crate::server::dispatch::{RequestAction, DEFAULT_REQUEST_TIMEOUT};
+pub use crate::server::dispatch::{RequestAction, SentRequest,
+                                  DEFAULT_REQUEST_TIMEOUT};
 pub use crate::server::io::{MessageReader, Output};
 use crate::server::io::{StdioMsgReader, StdioOutput};
-use crate::server::message::RawMessage;
+use crate::server::message::{RawMessage, RawMessageOrResponse, RawResponse};
 pub use crate::server::message::{
     Ack, BlockingNotificationAction, BlockingRequestAction,
     NoResponse, Notification, Request,
@@ -88,7 +89,7 @@ impl BlockingRequestAction for ShutdownRequest {
     fn handle<O: Output>(
         _id: RequestId,
         _params: Self::Params,
-         ctx: &mut ActionContext,
+         ctx: &mut ActionContext<O>,
         _out: O,
     ) -> Result<Self::Response, ResponseError> {
         if let Ok(ctx) = ctx.inited() {
@@ -105,7 +106,7 @@ impl BlockingRequestAction for ShutdownRequest {
     }
 }
 
-pub(crate) fn maybe_notify_unknown_configs<O: Output>(out: &O, unknowns: &[String]) {
+pub(crate) fn maybe_notify_unknown_configs<O: Output>(_out: &O, unknowns: &[String]) {
     use std::fmt::Write;
     if unknowns.is_empty() {
         return;
@@ -116,10 +117,11 @@ pub(crate) fn maybe_notify_unknown_configs<O: Output>(out: &O, unknowns: &[Strin
         write!(msg, "{}`{}` ", if first { ' ' } else { ',' }, key).unwrap();
         first = false;
     }
-    out.notify(Notification::<ShowMessage>::new(ShowMessageParams {
-        typ: MessageType::WARNING,
-        message: msg,
-    }));
+    warn!("{}", msg);
+    // out.notify(Notification::<ShowMessage>::new(ShowMessageParams {
+    //     typ: MessageType::WARNING,
+    //     message: msg,
+    // }));
 }
 
 #[allow(dead_code)]
@@ -192,7 +194,7 @@ impl BlockingRequestAction for InitializeRequest {
     fn handle<O: Output>(
         id: RequestId,
         mut params: Self::Params,
-        ctx: &mut ActionContext,
+        ctx: &mut ActionContext<O>,
         out: O,
     ) -> Result<NoResponse, ResponseError> {
         let mut dups = std::collections::HashMap::new();
@@ -276,8 +278,8 @@ pub struct LsService<O: Output> {
     output: O,
     server_send: channel::Sender<ServerToHandle>,
     server_receive: channel::Receiver<ServerToHandle>,
-    ctx: ActionContext,
-    dispatcher: Dispatcher,
+    ctx: ActionContext<O>,
+    dispatcher: Dispatcher<O>,
 }
 
 impl<O: Output> LsService<O> {
@@ -326,12 +328,15 @@ impl<O: Output> LsService<O> {
                     }
                 };
                 trace!("Read a message `{}`", msg_string);
-                match RawMessage::try_parse(&msg_string) {
-                    Ok(Some(rm)) => {
+                match RawMessageOrResponse::try_parse(&msg_string) {
+                    Ok(RawMessageOrResponse::Message(rm)) => {
                         debug!("Parsed a message: {}", rm.method);
                         send.send(ServerToHandle::ClientMessage(rm)).ok();
                     },
-                    Ok(None) => (),
+                    Ok(RawMessageOrResponse::Response(rr)) => {
+                        debug!("Parsed a response: {}", rr.id);
+                        send.send(ServerToHandle::ClientResponse(rr)).ok();
+                    },
                     Err(e) => {
                         error!("parsing error, {:?}", e);
                         output.custom_failure(
@@ -365,6 +370,11 @@ impl<O: Output> LsService<O> {
                         ServerStateChange::Break { exit_code }
                         => return exit_code,
                     },
+                ServerToHandle::ClientResponse(resp) => {
+                    if let ActionContext::Init(ctx) = &mut self.ctx {
+                        ctx.handle_request_response(resp, &self.output);
+                    }
+                },
                 ServerToHandle::ExitCode(code) => return code,
                 ServerToHandle::IsolatedAnalysisDone(path, context,
                                                      requests) => {
@@ -619,6 +629,7 @@ impl<O: Output> LsService<O> {
 #[derive(PartialEq, Debug)]
 pub enum ServerToHandle {
     ClientMessage(RawMessage),
+    ClientResponse(RawResponse),
     ExitCode(i32),
     IsolatedAnalysisDone(CanonPath, Option<CanonPath>, Vec<PathBuf>),
     DeviceAnalysisDone(CanonPath),
@@ -647,7 +658,7 @@ fn experimental_caps() -> Value {
     }).unwrap()
 }
 
-fn server_caps(_ctx: &ActionContext) -> ServerCapabilities {
+fn server_caps<O: Output>(_ctx: &ActionContext<O>) -> ServerCapabilities {
     ServerCapabilities {
         call_hierarchy_provider: None,
         declaration_provider: Some(DeclarationCapability::Simple(true)),
@@ -797,14 +808,12 @@ mod test {
     /// Some clients send empty object params for void params requests (see issue #1038).
     #[test]
     fn parse_shutdown_object_params() {
-        let raw = RawMessage::try_parse(
+        let raw = RawMessageOrResponse::try_parse(
             r#"{"jsonrpc": "2.0", "id": 2, "method": "shutdown", "params": {}}"#,
-        )
-        .ok()
-        .and_then(|x| x)
-        .expect("raw parse failed");
+        ).unwrap();
+        let parsed = raw.as_message().unwrap();
 
         let _request: Request<ShutdownRequest> =
-            raw.parse_as_request().expect("Boring validation is happening");
+            parsed.parse_as_request().expect("Boring validation is happening");
     }
 }